refact: 로그아웃 버튼 컴포넌트화
